VMAX driver - Block revert to snapshot for replicated volumes
Attempting to restore a replicated volume to snapshot will result in an
error, as this operation requires the RDF relationship to be suspended
first. It would be safest to throw an exception and block this operation
on replicated volumes. The user could instead create a new volume from
the snapshot and attach that, instead of reverting the original volume.
This patch implements this change.

Change-Id: If7b1dc408375d724fdc9ce6a3e256ad37e7c946f
Closes-bug: 1777871