Clean up unnecessary test params in test_rbd
Unit test was added that copied a test using ddt for supplied
values passed in, but the new test did not used ddt. This just
moves the values of those parameters into the body of the test.
It keeps the variable names to help keep the intent of the
values clear.

Change-Id: I5148f9500b0578dae457a9b561def8568165fbd1