Fix quota error when deleting temporary volume
When deleting a volume, always use the admin context to fetch the
admin_metadata that indicates whether it's a temporary volume.

This fixes a bug where a non-admin user failed to retrieve the
admin_metadata, which caused the volume to not be treated as temporary.
This, in turn, caused quota usage to be incorrectly updated.

Closes-Bug: #1778774
Change-Id: I8e6b0c726dc6498b28795a1ea0520ef05e53f047
(cherry picked from commit 8d9f8629013bcb880fb3e33f03b274c10befdab0)