NEC driver: Reuse iSCSI LUNs of detached volumes.
Currently, NEC driver does not reuse iSCSI LUNs of detached volumes.
The driver always use the LUN of (max(LUNs used) + 1) to attach
a volume.

If multi-target mode of NEC Storage is enabled, the driver works
without the problem because only LUN0s are used in the multi-target
mode.

This patch changes the using LUN to min(LUNs not used) for reusing
a LUN of detached volume with multi-target mode disabled.

Change-Id: I067cb09928c4c9d4ebb7159950fff526ee70f1bb
Closes-Bug: #1778669