Enable running py36 unit tests
This adds a tox setting for skipping default targets if the
specified version is not available. This allows tests to run
using python 3.6 if available, but conversely also allows
skipping python 3.5 or even 2.7 if those runtimes are not
present.

This only affects local default tox runs and does not change
what is tested in the gate.

Change-Id: Ia34fd32ab8d4672a2725468ee3d37697f51d5b3a