Add sphinx-feature-classification based support matrix
Cinder has always used a wiki page as the source of our
support matrix.  Unfortunately the wiki gives us no way
to ensure that the information is accurate and makes it
harder to track changes.  Moving to using the
sphinx-feature-classification library solves some of these
problems.

  *  It provides a programatic way to document driver support.
  *  It allows us to ensure that documentation is updated with
     changes.
  *  It will provide a snapshot of the state of driver support
     for a release at the time of release.

This matrix will serve as the truth for Cinder's driver support.
The existing wiki will be kept for historical purposes but I will
make a pointer to this new documentation and indicate that no
changes should be made to the wiki in the future.

Change-Id: I7c1b7fb539a48ec3b79e86c44ffe2d3005aeba25