Avoid os.getpgrp, breaking Windows c-backup support
A recent change introduced a regression, breaking Cinder Backup
support for Windows. It uses os.getpgrp, which isn't available on
Windows, using the process group as part of a lock name.

This lock ensures that only one of the cinder backup workers will
cleanup incomplete backups. It should be safe to just drop the
process group from the lock name, fixing this issue.

Closes-Bug: #1781168

Change-Id: I01edb5cd757a08b1b371e20ce5fa42fee888bbfb