Prepare Google Backup driver for latest libraries
With every new google library, there are small changes that need to be
done to the Google Backup Cinder driver to stay compatible.

This patch updatest the code and the unit tests for libraries:

  google-auth===1.5.0
  google-auth-httplib2===0.0.3
  google-api-python-client===1.7.3

While maintaining compatibility with earlier versions:

  google-auth===1.4.1
  google-auth-httplib2===0.0.3
  google-api-python-client===1.6.7

Code has also been manually tested against Google Cloud Storage to
ensure backup and restore works with the new libraries.

Change-Id: I4e0c36217338fd757445aa172f6e962fb9ec507c