INFINIDAT: validate_connector should validate by protocol
The INFINIDAT driver inherits from SanISCSIDriver for iSCSI-specific
members, but its validate_connector method checks for the
'initiator' property in 'connector', which is not always available
when using the FC protocol. Since the INFINIDAT driver supports
both protocols, the validate_connector method must be
overridden to check for the correct property in 'connector'.
This fixes an issue with IBM PowerVC.

Change-Id: Ia537f617baf72480b9c4536bdb7bc017784955e8