Use Tempest scenario job to run all scenario tests
Tempest has migrated the scenario job in Tempest
tree and renamed it to "tempest-scenario-all".[1]

This job runs all the Tempest scenario tests including
slow tests.

Old job running Tempest scenario test is running as experimental
in cinder[2]. That job will be removed in favor of "tempest-scenario-all"

This new job "tempest-scenario-all" will run as voting on check pipeline
where old job was running as experimental. This is fine as we want to run
all scenario tests on nova to avoid any regression.

Tempest also run "tempest-scenario-all" as voting job.

Detailed discussion on ML[3].

[1] http://git.openstack.org/cgit/openstack/tempest/tree/.zuul.yaml#n147
[2] https://git.openstack.org/cgit/openstack-infra/project-config/tree/zuul.d/projects.yaml#n3568
[3] http://lists.openstack.org/pipermail/openstack-dev/2018-May/130394.html

Change-Id: I984c9c449303effa5a3d8b754a3f4363add3dada