Non Windows per service lock for Backup service
The introduction of multiple processes on the Backup service in commit
66b1a0049ab78754a5e90c182b09ec501596458c broke the service in Windows.

On commit 1710839a4336dea996910b0894d69d1075c5954c this was fixed by
removing the original usage of os.getpgrp and using a global lock for
all backup services running on the host.

This global lock introduces an unnecessary bottleneck on non Windows
systems.

This patch fixes this bottleneck on non Windows systems by going back to
using a lock using the os.getpgrp method for non Windows systems and
using a global lock for Windows systems.

Related-Bug: #1781168
Change-Id: If7178695bf8ab751bcede14aa56d8becee60e17b