Update delete group snapshot API exception handling
When a role doesn't have access to delete group snapshot API, the
response code and exception that is returned is 400 BadRequest. This
is incorrect as a 403 PolicyNotAuthorized should be thrown when a
role doesn't have access to an API.

The HTTPBadRequest exception is raised here [0] regardless of what
the real exception is.

[0] https://github.com/openstack/cinder/blob/master/cinder/api/v3/group_snapshots.py#L93

Change-Id: I4a24cd35bbbba42ec3e360ce65fe278edc20d4bb
Closes-Bug: #1783640