Save the backup.service just before _run_backup
When the judgement of self.is_working() is false, it means the
backup service is not ready and the backup will be interrupted.
We will not allow to delete this backup util the backup service
is configured correctly and starts successfully.
This patch fix this, if the backup service was not working or we
just used a wrong backup service at that time, and now we want to
cleanup the backups, it is OK to do backup delete. There is no
need to do more!

Change-Id: I7163f026a497db8876c495d15bef9c006eadd17b