Update volume create schema to allow 'null'
When creating a volume from a source snapshot, size is not a required
parameter since it will be defaulted to the snapshot's original size. We
set it as not required in 8152f6c0cc0950fba33a67e62bf37e9ded7e1424, but
missed that the jsonschema validation also needed to be updated to allow
for the empty value.

Change-Id: I27b94c5579d5215fdd0fbd6c421aa5f6d9139bd8
Closes-bug: #1784821
Signed-off-by: Sean McGinnis <sean.mcginnis@gmail.com>