Add ability to specify minimum value when using capacity based QoS
When using capacity based QoS, it is possible that a user provisions
a volume that is very small but it unusable because of the very small
calculated IOPs.

This patch adds a few options to allow operators to allow for a minimum
value of IOPs, where the greater of the calculated value and minimum
is used as the QoS value.

Change-Id: I0fac45847f991d1f81ddcdd8674f8634f1cc4943