Correct volume-transfers API endpoint
Routing was set up to map to volume_transfers, but the cinder client and
other places expected it to be volume-transfer. The - is more correct
for API endpoints, and the client is currently in feature freeze, so
easiest way to correct this for this new feature is to update the
endpoint to the expected name.

Change-Id: Ib0002c441a2e860e8b74a4d9e6df890b398210f4
Closes-bug: #1785325
Signed-off-by: Sean McGinnis <sean.mcginnis@gmail.com>