Enable split logging for cinder-keystoneclient interaction
To enable split logging functionality[1], 'split_loggers' conf option
can be passed in below two ways:

1. As 'split_loggers' is already registred in [default] and [service_user]
group of cinder.conf we can use it while creating keystoneclient.
2. Register 'split_loggers' conf option in keystonemiddleware[2] so that we
can get it in [keystone_authtoken] group of cinder.conf and use it while
creating keystoneclient.

In this patch we are passing it from [service_user] group at the time of
creating keystoneclient.

[1]: https://review.openstack.org/#/c/505764/
[2]: https://review.openstack.org/#/c/578008/

Change-Id: Ic4972c6ebcac7115133ff7e906f756e68588a7b5