Allow extra parameters in volume_type_encryption
Change I6488cc89e60a31605ddf80c7aa56ddf4871d3c1a added schema validation
for volume type encryption, but it locked things down in a non-backwards
compatible way by not allowing extra parameters. This has caused
problems for at least Horizon. If we are going change this to not be
backwards compatible it should include a microversion bump and moving
the endpoint out of contrib to a standard endpoint.

To keep some schema validation at least, this just allows extra
parameters so existing API consumers should be unaffected. We can
address locking things down better in a future release when we have time
to do it right.

Closes-bug: #1783467
Closes-bug: #1786054

Change-Id: I45c254eca69471509e7ae4c957df51282e2d1776
Signed-off-by: Sean McGinnis <sean.mcginnis@gmail.com>