ScaleIO: Deprecate sio_allow_non_padded_thick_volumes
The config option sio_allow_non_padded_thick_volumes was added to
allow sites to continue to have non-padded thick volumes after it was
disabled by default due to security concerns. At the time it was only
thought to be an issue with thick volumes, but has since been determined
to also be an issue with thin.

To address the concerns with thin, the general option
sio_allow_non_padded__volumes was added, making the thick-only option
redundant. This new option was added without the deprecation of the
thick-only option to ease backporting to address the securty concern,
but we now want to work towards removing the thick option so we can just
have the one option for both provisioning types.

Related-bug: #1784871

Change-Id: Iaf7173cbcd9fc0929dabe3b1cb2db4b47c8bf0bd
Signed-off-by: Sean McGinnis <sean.mcginnis@gmail.com>