Modify Pure driver to configure PG/Pod names
Add two config options to the Pure Storage driver to allow setting the
names for replication PGs and Pods. This is beneficial in multi-backend
scenarios so that each backend can use a different PG/Pod. It also
allows for the use of existing PGs/Pods.

These options should not be modified after volumes exist on a backend.

Change-Id: Ie0dbf00fb2ecea3344e58d2b84fd4dfd6ae1daee