Remove auth_uri usage
auth_uri in keystonemiddleware was deprecated and will be removed
in Stein.

This patch changed it to auth_url or www_authenticate_uri correctly

Backporting to rocky as the option was deprecated there but our
documentation was still instructing that it be set during installation.

This modifies the cherry-pick to only bring back the documentation
changes. Those should have been updated so deployers started using the
preferred config option. Unfortunately, the doc changes were done along
with code changes that would make existing configs incompatible.
Following the documentation as it is now in the stable branch results in
deprecation warnings in the logs.

Conflict:
    cinder/quota_utils.py
    cinder/test.py
    cinder/tests/unit/contrib/test_quotas.py
    cinder/tests/unit/conf_fixture.py
    cinder/tests/unit/test_quota.py
    cinder/tests/unit/test_qutoa_utils.py

Closes-bug: #1823324
Change-Id: I754f6b27b3ebb87d4479aa366c9ee04af34fcaab
(cherry picked from commit 19dd18b17f2fbcca37b0c6633f47c0b7cdace9a7)