Avoid using dict.get() in assertions
The method dict.get(key) will return None in the following two cases:
A. if there is no key in the dictionary.
B. if the real value of the key is just None.

The above two cases will results in passing the assertion
which is not the expected behaviour of the test.

Change-Id: I634dca85b378ea42bd2e49aeb12aadea59a929fd
Related-Bug: #1613154