Add comment to online data migration command
One of the requirements of the online data migration command is that
they must be able to be run without any Cinder service running in order
to support a normal upgrade or a Fast-Forward upgrade.

With the introduction of the shared_targets feature we added an online
migration that requires the Cinder services to be running (it does RPC
calls).

To prevent this from happening again this patch adds a comment right
above the online_migrations tupple as a reminder for reviewers.

Change-Id: I911ddee5a62edc95bedbfd644804abd8b206e957