api-ref: add preconditions for volume transfer APIs
This adds the simple pre-conditions for the volume transfer
API (in its various forms) to the API reference. This came
up during discussions about nova server instance ownership
transfer and questioning whether or not in-use volumes could
be transferred and finding the API did not mention that case
at all.

Change-Id: I0bd42dd6a0e291979fb37bcab3823dfde369e7a9