Propose example volume protection tests
During the Stein PTG there was a session to discuss ways to make it
easier to ensure we can change policies without breakages. Ideally,
having a series of test classes than ensure the policy being using to
protect an API is exercised using the oslo.context and olso.policy
libraries.

This commit attempts to build out a set of example test cases that can
be used to increase test coverage in this area.

Change-Id: I88354ec57849e59764b1f5b0233be4fa1ef70c38