api-ref: clarify volume_type param in volume create API
The volume_type parameter to the volume create API can be
either the volume type name or ID, the cinder-api will look
it up either way. This wasn't obvious from the API reference
and required digging into the code to figure it out. This change
simply clarifies the usage in the API reference.

Change-Id: Iae8c43374dee9767e17252279822ccd2121d8246