Fix wrong NotFound in get_by_name_or_id
When we create or retype a volume using the volume type which
has a uuid format type name, the operation would be failed and
raise a 404 error.

There are a uuid check in "get_by_name_or_id()", and the
uuid-like type name would be mistakenly recognized as a id,
will use "get_volume_type"(by_id) to get volume type .

So, this patch try to fix this error, if we can't find a
type by uuid, we need call "get_volume_type_by_name" to check
again if we can get this type by name.

Co-Authored-By: Yikun Jiang <yikunkero@gmail.com>

Closes-Bug: #1794237
Change-Id: I83f12a5b858aced117286e8d2dee717d14dbdba6