Exclude disabled API versions from listing
The code for listing supported API versions would return a static list,
whether those versions were enabled or not. This changes the logic to
drop versions that are not enabled so we don't report back support for
something that then would cause a failure if the user tried to use it.

Closes-bug: #1788039

Change-Id: I52c1264b3b67a5c1ea00de1c0de5bea04c3096ee
Co-Authored-By: iain MacDonnell <iain.macdonnell@oracle.com>
Signed-off-by: Sean McGinnis <sean.mcginnis@gmail.com>