Fix wrong uuid recognized when create group
We can't create a group with a uuid format name volume type,
there is a uuid check in "volume_types_get_by_name_or_id()",
and the uuid-like type name would be mistakenly recognized as
a id, finally, it will use "get_volume_type"(by_id) to get
volume type and cause a 404 error.

So, this patch try to fix this error, if we can't find a
type by uuid, we need call "_volume_type_get_by_name" to check
again can we get this type by name.

Change-Id: Id09230bffc0ad83093bb6254b2e09aca5d1c58b1
Closes-bug: #1794716
Related-bug: #1794237