VNX: delete the LUN from VNX backend
Async migration is used in the process of creating a volume from a
snapshot and an internal temp snapshot is created.

Because the temp snapshot isn't deleted, the LUN hosting the temp
snapshot cannot be deleted from the VNX storage.

For example, a new volume, V2, is created from snapshot S1 of volume V1.
An internal temp snapshot S2 is created from copying S1. V1 now has two
snapshots, S1 and S2. Although we delete V1, V2 and S1 from Cinder,
S2 which is Cinder user-invisible isn't deleted and which causes V1 left
on VNX too.

The fix makes sure the snapshot S2 is deleted. Then the delay deletion
on V1 can be executed successfully.

Change-Id: Ib86729488ebfb0aea5d5d4f815a64e00258040e7
Closes-bug: #1794646