Fix unexpected behavior in _clone_image_volume
Current, if we upload a snapshot-volume to image, then create
a new volume from this image, this new volume is created from
original snapshot rather than snapshot-volume.

That means if we already make some data change on a
snapshot-volume, then we export this snapshot-volume to image,
this image will lose all change data.

This patch try to fix this, we skip the "snapshot_id" and
"source_volid" when we create tmp image volume.

Closes-bug: 1795098

Change-Id: Icf0dadc1ce4d008ff437ca5dcde1817d2e560e44