SF: Handle qos values on extend volume
For the Solidfire cinder driver when volume extend is called on a volume
that has a qos with scaledIOPs=True, the values scaleMin, scaleMax,
scaleBurst are ignored, thus when a volume grows the iops values stay
the same when they should either grow or shrink.

To correct the bug _retrieve_qos_setting has a new argument size with a
default value of zero. This allows all previous calls to function
properly, and saved duplicating code.

Extend volume now calls _retrive_qos_settings with the new size argument
and then includes that qos with the call to solidfire api  ModifyVolume.

Closes-Bug: #1783588
Change-Id: I8cb7b1eccbc63ef4077e86324afdd1a11409343d