Mark Veritas HyperScale Driver Unsupported
The Veritas HyperScale Driver 3rd Pary CI hasn't
reported in 153 days.  This does not meet the requirements
for Cinder's 3rd Party CI requirements.

As a result the driver is being marked unsupported in the
Stein release and should be removed in the 'T' release if
the problems are not resolved.

Change-Id: I4e20cdc555e9fb53dbc831f4464ccf80d9f01e7b