Handle drivers that do not support list manageable
Listing manageable volumes and snapshots is not a required feature for
drivers and not all of them support it. If a user attempts to list
objects when using a backend that does not support it, they currently
get a somewhat cryptic and internally-exposing error message:

AttributeError: 'RBDDriver' object has no attribute 'get_manageable_snapshots'

This updates the manager to better handle the case where a driver does
not support listing volumes and snapshots.

Closes-bug: #1796784

Change-Id: Ib6cdbe22ad22841c7a166f722eee981913396af4
Signed-off-by: Sean McGinnis <sean.mcginnis@gmail.com>
(cherry picked from commit b07c91abda47abb3432759f8a68625ce2585f823)