Follow Up: Generalizing is_replicated_str to is_boolean_str
Addressing the comment on the
commit Ic6c19414738367aa794522aaa0c1a963abcac723, the method
is_replicated_str consists the functionality of checking if the
string is valid boolean type or not which is being used for purposes
other than checking for replicated string.
This patch generalizes the name of the method.

Change-Id: Id299aeeea06a54d97fef2d8b0cabb83b7a518e19