Fix for cinder-manage db purge
When backend database is postgresql, "cinder-manage db purge" will throw
an exception. The root cause is that "deleted" column in table
quality_of_service_specs is of type boolean, but it is compared with an
integer. In mysql, there's no such issue because boolean in mysql is
tinyint(1).

Some logs:
DBAPIError exception wrapped from (psycopg2.ProgrammingError) operator
does not exist: boolean = integer

Closes-Bug: #1797048

Change-Id: I05b0edef4cc56518e99622599f01dbd6280a7f1b
Signed-off-by: Liang Fang <liang.a.fang@intel.com>