VNX: Add constraints for async migration
Async migration was introduced to accelerate the cloning of volume (
The cloned volume can be accessed after the migration starts, no need
to wait it for finishing).

However, before the migration finishes, snapshots cannot be created from
the source volume any more, which could affect subsequent clones from
the same source volume. That is the constraint of async migration.

The fix is adding the constraint to the doc of VNX driver. And give
users some advises on how to make the subsequent volume clone work.

Change-Id: I6ad3c6b967d5a33966b40c3a54c0f110576c8d87
Closes-bug: #1796825