Remove volume:get policy from test policy file
There is a policy file maintained specifically for testing that
overrides the default policies so that they are easier to test
against. Ideally, we should make it so that it easier to test the
default policies instead of overriding them since this would help us
prevent regression when policies are change. An example of how to do
this was implemented in the following patch:

  I88354ec57849e59764b1f5b0233be4fa1ef70c38

The more tests cases that are covered like that, the more policies we
can remove from cinder/tests/unit/policy.json. Once that file is empty
we should have satisfactory test coverage of the default policies that
are maintained in code.

Change-Id: Ia6c2e51fa88811e04f1c312f003b7b2524d90072