Forbidden to revert volume to a different size snapshot
As mentioned from original design:
"As we support extend volumes at present, we could meet the
case that snapshot is smaller than volume when reverting to
snapshot and it's obviously safe to revert in that case. But we
will still restrict to revert the volume which current volume size
is equal to snapshot, cause we don't support shrink volume yet
(that ability will be used in the generic method, if the driver
don't support revert to snapshot)."

So, we need add a size check before we revert a volume to a
snapshot. If the volume size is not equal to its latest
snapshot size, it will be raise a HTTPBadRequest(400).

Change-Id: Ib3cc95a10870822b9f597d66007699fb40908b61
Closes-bug: #1798503