Increase the length of resource property in quota_usages
When updating the volume type, the length of the name is checked to be
0-255. If the user input length is 255 characters
(e.g. volume_type_name='X*255'), when synchronizing the quota resource,
the resource name needs to be updated to the form of 'resource +
volume_type_name' in quota_usages [1], because the resource attribute in
the DB is set to 'String(255)', the length limit is exceeded.

Therefore, the resource column attribute in the quota_usages database table
needs to be changed from 'String(255)' to 'String(300)'.

[1]https://github.com/openstack/cinder/blob/master/cinder/db/sqlalchemy/api.py#L352

Closes-Bug: #1798327
Closes-Bug: #1608849
Change-Id: I6c30a6be750f6b9ecff7399dbb0aea66cdc097da