NetApp SolidFire: Thin provisioning scheduler support
Thin provisioning[1] is a feature supported by SolidFire hardware but
in the current driver implementation it is not managed by Cinder.
Therefore, scheduler options to control thin provisioning would not
work in this driver.

In this patch, we make the SolidFire driver compliant with Cinder thin
provisioning framework and also fix a few values that were not being
properly reported.

To keep the driver behavior consistent and avoid breaking rolling
upgrades, a new option is added. So, by default, the driver will
behave in the old way, and in the next releases this new option will
be the new default.

[1] https://specs.openstack.org/openstack/cinder-specs/specs/kilo/over-subscription-in-thin-provisioning.html

Change-Id: Ibba1dcf3deef27a9e1272d4e806f3e09ebd4ca4a
Closes-bug: #1784042