Allow using forward slashes in metadata
The recent changes introduced in Rocky disallowed the use of forward
slashes inside metadata which has caused software that relies on
metadata and uses those forward slashes stop functioning, a notable
example would be the Kubernetes cloud provider.

This patch allows the usage of a forward slash in the metadata as
well as extra specs, it is relatively harmless, unbreaks downstream
users and we already accept a range of symbols in there anyways.

Change-Id: Ibe180949205ef6985dc5c19904c127cbca361d53
Closes-Bug: #1798798