api-ref: cleanup status conditions for backup/restore APIs
The description for the backups API in both the v2 and v3
reference had a generic statement that only available volumes
can be backed up and restored, which is misleading since it
is possible to backup an in-use volume if the force parameter
is used.

This change drops the generic statement from the high-level
description and opts for per-API descriptions about the status
preconditions for backup and restore operations.

The v3 restore reference already mentioned the status precondition
so the only change to that is fixing the bullet list formatting.

Change-Id: Ib6c61af9ced2527884569fa15a848a215dd985b8
Closes-Bug: #1799499