Fix non iSCSI attach serialization
When the shared_targets concept was introduced it defaulted to True for
any driver not reporting that capability.  This meant that the
attachment of all volumes would be serialized even if they have no
concept of a "shared target".

This is caused because that feature was intended for iSCSI connections,
but was applied to all connection types, introducing an unnecessary
bottleneck on drivers like RBD or FC.

This patch makes sure to check that the storage_protocol is also iSCSI.

Closes-Bug: #1800136
Change-Id: I2805e8acf560cb941ddd3454477d89fe5a13d37f