VMAX Pike docs - no support for PowerMax OS
Although VMAX AFA can be upgraded to the PowerMax OS (5978),
we cannot support this configuration as both Solutions Enabler 9.0
and Unisphere for PowerMax 9.0 are required. These releases brought
about the deprecation of 'Workload', so if legacy Volume Types leveraged
'Workload' they will no longer work.  This only applies to Pike, we will
support this configuration in the Queens release and beyond.

Change-Id: Ie864195bd9cfd05d394a86ba0e1974036a762f39