Change the matching method of the backup driver
The backup driver name is configured uniformly from the Stein version.
For example, support for
'cinder.backup.drivers.service.ServiceBackupDriver' does not support
'cinder.backup.drivers.service'.

However, the above two backup drivers are supported in a stable version
before Rocky (including Rocky).

For the same storage backend, if the backup driver used to create the
backup file is 'cinder.backup.drivers.service.ServiceBackupDriver',
the backup driver configured when restoring the backup is
'cinder.backup.drivers.service'. The recovery of the backup file will
fail because the drivers do not match.

Therefore, the processing of this method needs to be compatible with the
backup driver for different configurations of the same backend.

The above 'service' is a proxy, it can be one of {'swift', 'nfs',
'ceph', 'glusterfs', 'google', 'tsm', 'posix'}

Closes-Bug: #1801316
Change-Id: I1d7ee395c89beda0e50cbe26b4abb4fc7749e15e