lvm: Avoid premature calls to terminate_connection for muiltiattach vols
Previously the LVM volume driver would always call termiante_connection
on the configured target driver regardless of of the associated volume
being attached to multiple instances on the same host. This isn't an
issue for the tgt target driver used by upstream CI as this call is a
noop but does cause the lioadm driver to prematurely remove specific
host ACLs for volumes that are still being used.

This change introduces a simple check to ensure that the target driver
is only called when a single attachment remains that is in turn using
the connector provided by the caller.

Finally as a result of this bugfix the changes introduced by
I84f607de13bc17b00609ad37121d8678f7f4a920 to disable the multiattach
feature when using the lioadm target driver are removed.

Closes-bug: #1786327
Change-Id: Ib5aa1b7578f7d3200185566ff5f8634dd519d020