Rollback the volume_types table when failed to update quota_usages
We create a new volume based on a volume type, then update the volume
type name to a new name of 255 characters in length.
When updating the volume type table, the quota_usages table is updated
synchronously. If DB error occurs in the quota_usages table, we need to
roll back the volume types table.Otherwise, when the volume is deleted,
the information of the quota_usages table will not be updated.

Closes-bug: #1802430
Related-Bug: #1608849

Change-Id: I040ccde7f04ae39258afd9b7322c939c8d796bde