api-ref: document encryption type provider choices
This updates the encrypted volume type provider
parameter descriptions for the create and update
requests to note the two choices of 'luks' and 'plain'.
Since we don't want those choices in the response
parameter descriptions, the parameters have to be
broken down a bit in the parameters.yaml file to
be request/response and optional/required.

There are currently other supported values for the
encryption type provider, including the legacy
classpath values, but we want people using the
non-legacy values so only luks and plain are
documented here, and luks is marked as the
recommended choice over plain (dm-crypt).

Change-Id: I496d30ecf6390f51cdd8e547b7345b3da30e92de
Closes-Bug: #1803513