An alternate way to fix retype notifier test case
When we do backup for a volume, the backup notification is called
periodically, but if we do other operations (like retype operation)
in parallel, there is a possible to record the "volume.retype" and
"backup.createprogress".

This patch try to mock the send notification methods to ensure this
doesn't happen.

Cloeses-Bug: #1803648

Co-Authored-By: Rajat Dhasmana <rajatdhasmana@gmail.com>

Change-Id: I4c44cdd6a4a11a3b7b5c57ff5aac19a8d01ef124
(cherry picked from commit 593d8dfa79980db651b88957f766604240e64e76)